Move syncing tracking code out of router into separate component
Renders an empty div to the page with some classes to indicate sync state to allow for easier tracking in functional tests, and potentially a starting point for future UI as well.
